requirements.txt
annotated-types==0.7.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
defusedxml==0.7.1
docstring_parser==0.16
firebase-admin==6.5.0
Flask==3.0.3
fonttools==4.53.1
fpdf==1.7.2
fpdf2==2.7.9
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
msgpack==1.0.8
numpy==2.0.0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
reportlab==4.2.2
requests==2.32.3
rsa==4.9
shapely==2.0.5
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
vertexai==1.49.0
Werkzeug==3.0.3
